pyproject.toml
[project]
name = "investsight"
version = "0.1.0"
description = "Crypto and stock portfolio tracker"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django>=5.0,<6.0",
    "django-environ>=0.11",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "requests>=2.31",
    "yfinance>=0.2",
    "tenacity>=8.2",
    "structlog>=24.0",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-mock>=3.12",
    "responses>=0.25",
    "model-bakery>=1.18",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"
